+++ pyngrok/conf.py
@@ -27,9 +27,9 @@ class PyngrokConfig:
     :var keep_thread_alive: Whether or not the `ngrok` process should continue to be monitored in a separate thread
         after it has finished starting.
     :vartype keep_thread_alive: bool
-    :var log_func: A callback that will be invoked each time `ngrok` emits a log. `keep_thread_alive` must be
+    :var log_event_callback: A callback that will be invoked each time `ngrok` emits a log. `keep_thread_alive` must be
         True for the function to continue to be called after `ngrok` finishes starting.
-    :vartype log_func: function
+    :vartype log_event_callback: function
     :var boot_timeout: The max number of seconds to wait for `ngrok` to start before timing out.
     :vartype boot_timeout: int
     :var max_logs: The max number of logs to store in NgrokProcess's logs variable.
@@ -44,16 +44,16 @@ class PyngrokConfig:
                  auth_token=None,
                  region=None,
                  keep_thread_alive=True,
-                 log_func=None,
+                 log_event_callback=None,
                  boot_timeout=15,
-                 max_logs=500,
+                 max_logs=100,
                  request_timeout=4):
         self.ngrok_path = DEFAULT_NGROK_PATH if ngrok_path is None else ngrok_path
         self.config_path = DEFAULT_CONFIG_PATH if config_path is None else config_path
         self.auth_token = auth_token
         self.region = region
         self.keep_thread_alive = keep_thread_alive
-        self.log_func = log_func
+        self.log_event_callback = log_event_callback
         self.boot_timeout = boot_timeout
         self.max_logs = max_logs
         self.request_timeout = request_timeout